requirements.txt
Flask==2.3.3
Pillow==10.0.1
pytesseract==0.3.10
# dlib
# face-recognition==1.3.0
pyspellchecker==0.7.2
langdetect==1.0.9
